pyproject.toml
[project]
name = "video-uploader-fullstack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.135.1",
    "fastapi-users[sqlalchemy]>=15.0.4",
    "imagekitio<5",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "uvicorn[standard]>=0.41.0",
]
